remove time_manager and include it into gui_manager
