+++ gui_manager.py
@@ -1,7 +1,14 @@
 import tkinter
 from tkinter import Tk, Label, Button, Frame, Misc, Text, messagebox
-from time_manager import Chrono
 
+
+class Chrono:
+
+    def __init__(self, is_pause=False):
+        self.seconds = 0
+        self.process = None
+        self.is_pause = is_pause
+        self.max_time = 25 * 60
 
 class Gui:
